+++ testing/main_testing.cpp
@@ -16,8 +16,8 @@ int main(int argc, char **argv)
 {
     ::testing::InitGoogleTest(&argc, argv);
 
-    unsigned long long int handler = almacenamiento::IAdministradorAlmacenamiento::iniciarNuevo("config_testing.json");
-    unsigned long long int handler_2 = almacenamiento::IAdministradorAlmacenamiento::iniciarNuevo("config_testing_2.json");
+    uintmax_t handler = almacenamiento::IAdministradorAlmacenamiento::iniciarNuevo("config_testing.json");
+    uintmax_t handler_2 = almacenamiento::IAdministradorAlmacenamiento::iniciarNuevo("config_testing_2.json");
 
     almacenamiento::IAdministradorAlmacenamiento::getInstancia(handler)->abrir();
     almacenamiento::IAdministradorAlmacenamiento::getInstancia(handler_2)->abrir();
